excel files in test dir added to repository

+++ tests/test_eu_state_aids.py
@@ -17,6 +17,7 @@ runner = CliRunner()
 encoded_test_year = "8U%2BIPGXBzzM%3D"  # 2015 encoded
 excel_test_url = f"http://2020.eufunds.bg/en/0/0/Project/ExportToExcel?StFrom={encoded_test_year}&" \
     f"StTo={encoded_test_year}&ShowRes=True&IsProgrammeSelected=False&IsRegionSelected=False"
+state_aids_url = "https://stateaid.minfin.bg/document/860"
 local_test_path = Path("./data/test/")
 
 
@@ -55,7 +56,8 @@ def test_non_existing_command():
 def test_bg_fetch():
 
     with requests_mock.Mocker() as mock:
-        shutil.rmtree(local_test_path)
+        if os.path.exists(local_test_path):
+            shutil.rmtree(local_test_path)
 
         mock.get(excel_test_url, text='resp')
         result = runner.invoke(
@@ -74,11 +76,16 @@ def test_bg_export():
 
     with open(Path("./tests") / f"bg_projects_sample.xlsx", mode='rb') as sample_xls:
         sample_content = sample_xls.read()
+    with open(Path("./tests") / f"bg_state_aids.xlsx", mode='rb') as state_aids_xls:
+        state_aids_content = state_aids_xls.read()
 
     with requests_mock.Mocker() as mock:
-        shutil.rmtree(local_test_path)
+        if os.path.exists(local_test_path):
+            shutil.rmtree(local_test_path)
 
-        mock.get(excel_test_url, text=str(sample_content))
+        mock.get(excel_test_url, content=sample_content)
+        mock.get(state_aids_url, content=state_aids_content)
+
         result = runner.invoke(
             app, ["bg", "fetch", "2015", "--local-path=./data/test"], prog_name='eu-state-aids'
         )
@@ -94,4 +101,9 @@ def test_bg_export():
             app, ["bg", "export", "2015", "--local-path=./data/test"], prog_name='eu-state-aids'
         )
         assert result.exit_code == 0
-        assert result.stdout == "Fetching EU data for year: 2015\nFile saved to data/test/projects_2015.xlsx\n"
+        assert 'DataFrame with 470 rows created from data/test/projects_2015.xlsx' in result.stdout
+        assert '190 matches found.' in result.stdout
+
+        assert(os.path.exists(local_test_path / "2015.csv"))
+        with open(local_test_path / "2015.csv", mode='r') as csv:
+            assert(len(csv.readlines()) == 191)
